refactor split domains parser
